refactor electrum queries
