+++ btc-backend/src/routes/electrum/mod.rs
@@ -1,0 +1,7 @@
+pub mod electrum_query;
+pub mod electrum_peers;
+pub mod electrum_servers;
+
+pub use electrum_query::electrum_query;
+pub use electrum_peers::electrum_peers;
+pub use electrum_servers::electrum_servers;
