+++ btc-backend/src/routes/mod.rs
@@ -1,10 +1,3 @@
 pub mod api_info;
 pub mod health;
 pub mod electrum;
-
-#[allow(unused_imports)]
-pub use api_info::api_info;
-#[allow(unused_imports)]
-pub use health::health_check;
-#[allow(unused_imports)]
-pub use electrum::electrum_peers;
